+++ Modules/cjkcodecs/multibytecodec.c
@@ -4,6 +4,10 @@
  * Written by Hye-Shik Chang <perky@FreeBSD.org>
  */
 
+#ifndef Py_BUILD_CORE_BUILTIN
+#  define Py_BUILD_CORE_MODULE 1
+#endif
+
 #include "Python.h"
 #include "structmember.h"         // PyMemberDef
 #include "multibytecodec.h"
